test(axum): add server tests for axum ticket server
