+++ htsget-axum/src/server/ticket.rs
@@ -1,0 +1,292 @@
+use crate::error::Result;
+use crate::handlers::{get, post, reads_service_info, variants_service_info};
+use crate::server::{configure_cors, AppState, BindServer, Server};
+use axum::routing::get;
+use axum::Router;
+use htsget_config::config::cors::CorsConfig;
+use htsget_config::config::{ServiceInfo, TicketServerConfig};
+use htsget_search::HtsGet;
+use std::net::SocketAddr;
+use std::sync::Arc;
+use tower::ServiceBuilder;
+use tower_http::trace::TraceLayer;
+
+impl From<TicketServerConfig> for BindServer {
+  /// Returns a ticket server with TLS enabled if the tls config is not None or without TLS enabled
+  /// if it is None.
+  fn from(config: TicketServerConfig) -> Self {
+    let addr = config.addr();
+    let cors = config.cors().clone();
+
+    match config.into_tls() {
+      None => Self::new(addr, cors),
+      Some(tls) => Self::new_with_tls(addr, cors, tls),
+    }
+  }
+}
+
+/// An data block server.
+#[derive(Debug)]
+pub struct TicketServer<H> {
+  server: Server,
+  htsget: H,
+  service_info: ServiceInfo,
+  cors: CorsConfig,
+}
+
+impl<H> TicketServer<H>
+where
+  H: HtsGet + Clone + Send + Sync + 'static,
+{
+  /// Create a new ticket server.
+  pub fn new(server: Server, htsget: H, service_info: ServiceInfo, cors: CorsConfig) -> Self {
+    Self {
+      server,
+      htsget,
+      service_info,
+      cors,
+    }
+  }
+
+  /// Run the data server, using the key and certificate.
+  pub async fn serve(self) -> Result<()> {
+    self
+      .server
+      .serve(Self::router(self.htsget, self.service_info, self.cors))
+      .await
+  }
+
+  /// Create the router for the ticket server.
+  pub fn router(htsget: H, service_info: ServiceInfo, cors: CorsConfig) -> Router {
+    Router::default()
+      .route(
+        "/reads/service-info",
+        get(reads_service_info::<H>).post(reads_service_info::<H>),
+      )
+      .route("/reads/*id", get(get::reads).post(post::reads))
+      .route(
+        "/variants/service-info",
+        get(variants_service_info::<H>).post(variants_service_info::<H>),
+      )
+      .route("/variants/*id", get(get::variants).post(post::variants))
+      .layer(
+        ServiceBuilder::new()
+          .layer(TraceLayer::new_for_http())
+          .layer(configure_cors(cors)),
+      )
+      .with_state(AppState::new(Arc::new(htsget), service_info))
+  }
+
+  /// Get the local address the server has bound to.
+  pub fn local_addr(&self) -> Result<SocketAddr> {
+    self.server.local_addr()
+  }
+}
+
+#[cfg(test)]
+mod tests {
+  use std::convert::Infallible;
+  use std::path::Path;
+  use std::result;
+  use std::str::FromStr;
+
+  use super::*;
+  use async_trait::async_trait;
+  use axum::body::{to_bytes, Body};
+  use axum::response::Response;
+  use htsget_config::config::Config;
+  use htsget_config::types::JsonResponse;
+  use htsget_test::http::server::expected_url_path;
+  use htsget_test::http::{
+    config_with_tls, cors, default_test_config, server, Header, Response as TestResponse,
+    TestRequest, TestServer,
+  };
+  use http::header::HeaderName;
+  use http::Request;
+  use tempfile::TempDir;
+  use tower::ServiceExt;
+
+  struct AxumTestServer {
+    config: Config,
+  }
+
+  struct AxumTestRequest<T>(T);
+
+  impl TestRequest for AxumTestRequest<Request<Body>> {
+    fn insert_header(mut self, header: Header<impl Into<String>>) -> Self {
+      self.0.headers_mut().insert(
+        HeaderName::from_str(&header.name.into()).expect("expected valid header name"),
+        header
+          .value
+          .into()
+          .parse()
+          .expect("expected valid header value"),
+      );
+      self
+    }
+
+    fn set_payload(mut self, payload: impl Into<String>) -> Self {
+      *self.0.body_mut() = Body::new(payload.into());
+      self
+    }
+
+    fn uri(mut self, uri: impl Into<String>) -> Self {
+      let uri = uri.into();
+      *self.0.uri_mut() = uri.parse().expect("expected valid uri");
+      self
+    }
+
+    fn method(mut self, method: impl Into<String>) -> Self {
+      *self.0.method_mut() = method.into().parse().expect("expected valid method");
+      self
+    }
+  }
+
+  impl Default for AxumTestServer {
+    fn default() -> Self {
+      Self {
+        config: default_test_config(),
+      }
+    }
+  }
+
+  #[async_trait(? Send)]
+  impl TestServer<AxumTestRequest<Request<Body>>> for AxumTestServer {
+    async fn get_expected_path(&self) -> String {
+      let mut bind_data_server = BindServer::from(self.get_config().data_server().clone());
+      let server = bind_data_server
+        .bind_data_server("/data".to_string())
+        .await
+        .unwrap();
+      let addr = server.local_addr();
+
+      let path = self.get_config().data_server().local_path().to_path_buf();
+      tokio::spawn(async move { server.serve(path).await.unwrap() });
+
+      expected_url_path(self.get_config(), addr.unwrap())
+    }
+
+    fn get_config(&self) -> &Config {
+      &self.config
+    }
+
+    fn get_request(&self) -> AxumTestRequest<Request<Body>> {
+      AxumTestRequest(Request::default())
+    }
+
+    async fn test_server(
+      &self,
+      request: AxumTestRequest<Request<Body>>,
+      expected_path: String,
+    ) -> TestResponse {
+      let response = self.get_response(request.0).await.unwrap();
+
+      let status: u16 = response.status().into();
+      let headers = response.headers().clone();
+      let bytes = to_bytes(response.into_body(), usize::MAX)
+        .await
+        .unwrap()
+        .to_vec();
+
+      TestResponse::new(status, headers, bytes, expected_path)
+    }
+  }
+
+  impl AxumTestServer {
+    fn new_with_tls<P: AsRef<Path>>(path: P) -> Self {
+      Self {
+        config: config_with_tls(path),
+      }
+    }
+
+    async fn get_response(&self, request: Request<Body>) -> result::Result<Response, Infallible> {
+      let app = TicketServer::router(
+        self.config.clone().owned_resolvers(),
+        self.config.service_info().clone(),
+        self.config.ticket_server().cors().clone(),
+      );
+
+      app.oneshot(request).await
+    }
+  }
+
+  #[tokio::test]
+  async fn get_http_tickets() {
+    server::test_get::<JsonResponse, _>(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn post_http_tickets() {
+    server::test_post::<JsonResponse, _>(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_get_http_tickets() {
+    server::test_parameterized_get::<JsonResponse, _>(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_post_http_tickets() {
+    server::test_parameterized_post::<JsonResponse, _>(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_post_class_header_http_tickets() {
+    server::test_parameterized_post_class_header::<JsonResponse, _>(&AxumTestServer::default())
+      .await;
+  }
+
+  #[tokio::test]
+  async fn service_info() {
+    server::test_service_info(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn get_https_tickets() {
+    let base_path = TempDir::new().unwrap();
+    server::test_get::<JsonResponse, _>(&AxumTestServer::new_with_tls(base_path.path())).await;
+  }
+
+  #[tokio::test]
+  async fn post_https_tickets() {
+    let base_path = TempDir::new().unwrap();
+    server::test_post::<JsonResponse, _>(&AxumTestServer::new_with_tls(base_path.path())).await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_get_https_tickets() {
+    let base_path = TempDir::new().unwrap();
+    server::test_parameterized_get::<JsonResponse, _>(&AxumTestServer::new_with_tls(
+      base_path.path(),
+    ))
+    .await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_post_https_tickets() {
+    let base_path = TempDir::new().unwrap();
+    server::test_parameterized_post::<JsonResponse, _>(&AxumTestServer::new_with_tls(
+      base_path.path(),
+    ))
+    .await;
+  }
+
+  #[tokio::test]
+  async fn parameterized_post_class_header_https_tickets() {
+    let base_path = TempDir::new().unwrap();
+    server::test_parameterized_post_class_header::<JsonResponse, _>(&AxumTestServer::new_with_tls(
+      base_path.path(),
+    ))
+    .await;
+  }
+
+  #[tokio::test]
+  async fn cors_simple_request() {
+    cors::test_cors_simple_request(&AxumTestServer::default()).await;
+  }
+
+  #[tokio::test]
+  async fn cors_preflight_request() {
+    cors::test_cors_preflight_request(&AxumTestServer::default()).await;
+  }
+}
